chore: clean up outdated files
