+++ .gitignore
@@ -14,4 +14,5 @@
 # Dependency directories (remove the comment below to include it)
 # vendor/
 
-/host
+/charmil
+/hostCLI
